projects, removed project.repo, shifting it's logic into buildercore/checks.py. projects, removed the 'protocol,host,path' triple wrangling.
